move createUpdatedNavigation to util functions
